minor changes to 2 and 3, created 4
there was a location in the column headers were there was a space
rather than a tab which was throwing everything off so I corrected
that. Version 4 rearranges V1V3 9-12 from github. It looks good except
that the study number comes out as one or two digits rather than 3 (the
leading zero(s) is/are missing)