remove extra tabs form the end of lines in MiSeq19
When moving all of the other updated MiSeq mapping files over to the
Club_Grice GitHub folder I accidentally overwrote an updated version of
MiSeq19. I just copied and pasted the still un-synced (and up to date)
version of the file from the GitHub website into a text file and had to
remove some extra tabs from the end of each line before it matched the
rest of the mapping files. This script removes the extra tabs.